// src/models/selectors.ts
export enum Selectors {
    POPUP_DIALOG = "//*[@role='dialog']",
    FIRST_BUTTON_ON_POPUP_DIALOG = "//*[@role='dialog']//button[1]",
    USERNAME_INPUT = "//input[@name='username']",
    PASSWORD_INPUT = "//input[@name='password']",
    SUBMIT_BUTTON = "//button[@type='submit']",
    BUTTON_TYPED_BUTTON = "//button[@type='button']",
}